replace old Send component
